msm: clock-mdss: correct check for MDSS GDSC status
MDSS GDSC may be always on but clock is off in reset state.
Therefore, the correct way to check GDSC status is read both
gdsc enable bit and the clock reset bit. This change is
necessary to prevent potential crashes that may occur if
MDSS registers are accessed when MDSS GDSC is not enabled.

Change-Id: I501450beed3ea798329b34b96f2966cfefc901c2
Signed-off-by: Huaibin Yang <huaibiny@codeaurora.org>
Signed-off-by: Pawan Kumar <pavaku@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>
Signed-off-by: flar2 <asegaert@gmail.com>